+++ CHANGES.rst
@@ -49,6 +49,11 @@ Moved ``command_list_as_table()`` from `utils` into ``plans/command_list``.
 
 ``utils/``: Reorganized ``utils.py`` and ``_utils/`` into ``utils/`` subpackage.
 
+New Features and/or Enhancements
+---------------------------------------------
+
+* Add support for synApps ``sseq`` record and ``userStringSeq`` databases.
+
 Maintenance
 ---------------